Refactored the code
- Earlier all the files were present in a single folder for applications management feature, and it also contained the apps local database.
- Following changes were made -
  - `AppDatabase` is now moved into the `data` folder inside `common` folder.
  - The folder inside `data` folder, where all the models for tables are stored is renamed to `entity`.
  - All the files are changed due too subsequence refactoring.
