+++ app/src/main/java/com/studbudd/application_tracker/common/data/models/ApiResponse.kt
@@ -1,0 +1,15 @@
+package com.studbudd.application_tracker.common.data.models
+
+
+import com.squareup.moshi.Json
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
+data class ApiResponse<T>(
+    @Json(name = "data")
+    val data: T?,
+    @Json(name = "message")
+    val message: String,
+    @Json(name = "success")
+    val success: Boolean
+)
