+++ app/src/main/java/com/studbudd/application_tracker/di/AppModule.kt
@@ -10,8 +10,8 @@ import com.google.android.gms.auth.api.signin.GoogleSignInOptions
 import com.studbudd.application_tracker.BuildConfig
 import com.studbudd.application_tracker.common.domain.ClearAppDataUseCase
 import com.studbudd.application_tracker.common.domain.SharedPreferencesManager
-import com.studbudd.application_tracker.feature_applications_management.data.AppDatabase
-import com.studbudd.application_tracker.feature_applications_management.data.ApplicationsRepository
+import com.studbudd.application_tracker.common.data.AppDatabase
+import com.studbudd.application_tracker.feature_applications_management.data.repo.ApplicationsRepository
 import com.studbudd.application_tracker.feature_user.data.dao.AuthUserRemoteDao
 import com.studbudd.application_tracker.feature_user.data.dao.UserRemoteDao
 import com.studbudd.application_tracker.feature_user.data.repo.UserRepository
