+++ app/src/main/java/com/studbudd/application_tracker/feature_user/data/dao/AuthUserRemoteDao.kt
@@ -1,7 +1,7 @@
 package com.studbudd.application_tracker.feature_user.data.dao
 
-import com.studbudd.application_tracker.common.models.ApiResponse
-import com.studbudd.application_tracker.feature_user.data.models.UserRemote
+import com.studbudd.application_tracker.common.data.models.ApiResponse
+import com.studbudd.application_tracker.feature_user.data.entity.UserRemote
 import retrofit2.Response
 import retrofit2.http.GET
