+++ app/src/main/java/com/studbudd/application_tracker/feature_user/data/dao/UserRemoteDao.kt
@@ -1,12 +1,10 @@
 package com.studbudd.application_tracker.feature_user.data.dao
 
-import com.studbudd.application_tracker.common.models.ApiResponse
-import com.studbudd.application_tracker.feature_user.data.models.UserRemote
-import com.studbudd.application_tracker.feature_user.data.models.requests.LoginRequest
-import com.studbudd.application_tracker.feature_user.data.models.response.LoginResponse
+import com.studbudd.application_tracker.common.data.models.ApiResponse
+import com.studbudd.application_tracker.feature_user.data.entity.requests.LoginRequest
+import com.studbudd.application_tracker.feature_user.data.entity.response.LoginResponse
 import retrofit2.Response
 import retrofit2.http.Body
-import retrofit2.http.GET
 import retrofit2.http.POST
 
 interface UserRemoteDao {
