+++ app/src/main/java/com/studbudd/application_tracker/feature_user/data/entity/RemoteIdTuple.kt
@@ -1,0 +1,10 @@
+package com.studbudd.application_tracker.feature_user.data.entity
+
+data class RemoteIdTuple(
+    val remoteId: String?
+) {
+
+    fun isConnectedWithRemoteDatabase() =
+        remoteId != null
+
+}
