+++ app/src/main/java/com/studbudd/application_tracker/feature_user/data/entity/response/LoginResponse.kt
@@ -1,0 +1,12 @@
+package com.studbudd.application_tracker.feature_user.data.entity.response
+
+import com.squareup.moshi.Json
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
+data class LoginResponse(
+    @Json(name = "accessToken")
+    val accessToken: String,
+    @Json(name = "refreshToken")
+    val refreshToken: String,
+)
