--- app/src/main/java/com/studbudd/application_tracker/feature_user/data/models/RemoteIdTuple.kt
@@ -1,10 +1,0 @@
-package com.studbudd.application_tracker.feature_user.data.models
-
-data class RemoteIdTuple(
-    val remoteId: String?
-) {
-
-    fun isConnectedWithRemoteDatabase() =
-        remoteId != null
-
-}
