+++ app/src/main/java/com/studbudd/application_tracker/feature_user/data/repo/UserRepository.kt
@@ -1,9 +1,9 @@
 package com.studbudd.application_tracker.feature_user.data.repo
 
-import com.studbudd.application_tracker.common.models.Resource
-import com.studbudd.application_tracker.feature_user.data.models.UserLocal
-import com.studbudd.application_tracker.feature_user.data.models.UserRemote
-import com.studbudd.application_tracker.feature_user.data.models.response.LoginResponse
+import com.studbudd.application_tracker.common.data.models.Resource
+import com.studbudd.application_tracker.feature_user.data.entity.UserLocal
+import com.studbudd.application_tracker.feature_user.data.entity.UserRemote
+import com.studbudd.application_tracker.feature_user.data.entity.response.LoginResponse
 import kotlinx.coroutines.flow.Flow
 
 interface UserRepository {
