+++ app/src/main/java/com/studbudd/application_tracker/feature_user/domain/use_cases/CreateRemoteUserUseCase.kt
@@ -1,7 +1,7 @@
 package com.studbudd.application_tracker.feature_user.domain.use_cases
 
-import com.studbudd.application_tracker.common.models.Resource
-import com.studbudd.application_tracker.feature_user.data.models.UserLocal
+import com.studbudd.application_tracker.common.data.models.Resource
+import com.studbudd.application_tracker.feature_user.data.entity.UserLocal
 import com.studbudd.application_tracker.feature_user.data.repo.UserRepository
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
